pyproject.toml
[tool.poetry]
name = "async"
version = "0.1.0"
description = ""
authors = ["Marcus Braga <mvbraga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
uvicorn = "^0.20.0"
channels = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
